backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.5
langchain-openai==0.0.5
pydantic==2.5.3
requests==2.31.0
python-dotenv==1.0.0
